+++ CHANGELOG.md
@@ -7,6 +7,11 @@ Mapbox welcomes participation and contributions from everyone.
 ### Breaking changes ⚠️
 
   - `OrnamentOptions.logo._isVisible` and `OrnamentOptions.attributionButton._isVisible` have been replaced with `OrnamentOptions.logo.visibility` and `OrnamentOptions.attributionButton.visibility`. ([#326](https://github.com/mapbox/mapbox-maps-ios/pull/326))
+
+### Bug fixes 🐞
+
+  - Fixed an issue where location pucks would not be rendered. ([#331](https://github.com/mapbox/mapbox-maps-ios/pull/331))
+
 ## 10.0.0-beta.19 - May 6, 2021
 
 ### Breaking changes ⚠️